# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación para API REST de Recetas",
        "contact": {
            "email": "eescobar@cdhidalgo.tecnm.mx"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://apibecas2-production.up.railway.app"
        }
    ],
    "paths": {
...
        "/api/direcciones": {
            "get": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Obtener lista de direcciones",
                "operationId": "f8c29a712a77b019c23b33c2b12b7972",
                "responses": {
                    "200": {
                        "description": "Lista de direcciones recuperada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DireccionResource"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Crear una nueva dirección",
                "operationId": "ae54cb09d534263f9ecc99c7efe66ae5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_estudiante",
                                    "municipio",
                                    "colonia",
                                    "calle",
                                    "numero"
                                ],
                                "properties": {
                                    "id_estudiante": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "municipio": {
                                        "type": "string",
                                        "example": "Guadalajara"
                                    },
                                    "colonia": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "calle": {
                                        "type": "string",
                                        "example": "Av. Principal"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dirección creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DireccionResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/direcciones/{id}": {
            "get": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Obtener una dirección específica",
                "operationId": "23c0eb511913cd49190b950a1563935a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la dirección",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dirección encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DireccionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dirección no encontrada"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Actualizar una dirección existente",
                "operationId": "f97ba5736ebc2758aef2d4f19ea241c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la dirección a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_estudiante": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "municipio": {
                                        "type": "string",
                                        "example": "Guadalajara"
                                    },
                                    "colonia": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "calle": {
                                        "type": "string",
                                        "example": "Av. Principal"
                                    },
                                    "numero": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dirección actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DireccionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dirección no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Direcciones"
                ],
                "summary": "Eliminar una dirección",
                "operationId": "aa8cbac694ebf09f0c6bd7f37339c6b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la dirección a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dirección eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dirección eliminada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Dirección no encontrada"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "DireccionResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_estudiante": {
                        "type": "integer",
                        "example": 1
                    },
                    "municipio": {
                        "type": "string",
                        "example": "Guadalajara"
                    },
                    "colonia": {
                        "type": "string",
                        "example": "Centro"
                    },
                    "calle": {
                        "type": "string",
                        "example": "Av. Principal"
                    },
                    "numero": {
                        "type": "integer",
                        "example": 123
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum API token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}